pyproject.toml
[tool.poetry]
name = "movie_lib"
version = "0.1.0"
description = ""
authors = ["cyr <cyr@zolotarev.pp.ua>"]

[tool.poetry.dependencies]
python = "^3.8"
flask-restx = "^0.4.0"
gunicorn = "^20.1.0"
Flask-Login = "^0.5.0"
Flask-SQLAlchemy = "^2.4.4"
pytest = "^6.2.4"
pylint = "^2.8.3"
psycopg2-binary = "^2.9.1"
marshmallow = "^3.12.1"
marshmallow-sqlalchemy = "^0.26.1"
black = "^21.6b0"
Faker = "^8.9.1"
sqlalchemy-seeder = "^0.3.0"
Flask-DB = "^0.3.2"
data = "^0.4"
production = "^0.0.0"
short = "^1.0.6"
prod = "^0.7"

[tool.poetry.dev-dependencies]
pylint = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
